引数の順序変更と型安全性の向上
`GetIdxMouseOnRectEdge` 関数の引数の順序を変更し、`std::vector<LabelObj>& objs` を `EditMode& editMode` の前に移動しました。また、`for` ループのインデックス型を `int` から `size_t` に変更し、オーバーフローのリスクを軽減しました。さらに、`WndProc` 関数内でマウスカーソルの位置を取得する処理にコメントを追加し、コードの可読性を向上させました。

+++ CPP_Anotation6/CPP_AnnoFunctions.cpp
@@ -1028,11 +1028,11 @@ int GetIdxMouseOnRectEdge(
 ){
     int _idx=-1;
 
-    //全部
-    for (int i = 0; i < objs.size(); i++)
+	//全部の矩形のマウスオーバー状態を解除
+    for (size_t i = 0; i < objs.size(); i++)
         objs[i].mOver = false;
 
-    for (int i = 0; i < objs.size(); i++)
+    for (size_t i = 0; i < objs.size(); i++)
     {
         //辺の位置が返ってくる。
 		EditMode _em = IsMouseOnRectEdge(pt, objs[i], overlap);
